backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
playwright==1.41.2
pillow==10.2.0
websockets==12.0
python-multipart==0.0.7 